requirements.txt
python-telegram-bot[job-queue]>=21.3
httpx==0.27.0
python-dotenv==1.0.0
aiosqlite==0.19.0
